+++ src/scriptlets/log.js
@@ -5,6 +5,9 @@
  * A simple scriptlet which only purpose is to print arguments to console.
  * This scriptlet can be helpful for debugging and troubleshooting other scriptlets.
  *
+ * Related ABP source:
+ * https://gitlab.com/eyeo/snippets/-/blob/main/source/introspection/log.js
+ *
  * ### Examples
  *
  * ```adblock
@@ -16,4 +19,7 @@
 export function log(...args) {
     console.log(args); // eslint-disable-line no-console
 }
-log.names = ['log'];
+log.names = [
+    'log',
+    'abp-log',
+];